Fixes to breadcrumbs in admin theme.
git-svn-id: http://svn.cmsmadesimple.org/svn/cmsmadesimple/branches/1.6.x@5632 3d254a34-79dc-0310-9e5f-be208747d8a0

+++ admin/themes/default/defaultTheme.php
@@ -139,7 +139,8 @@ class defaultTheme extends AdminTheme
 			if ($counter > 0) {
 				echo " &#187; ";
 			}
-			if (str_replace('&amp;', '&', $crumb['url']) != basename($_SERVER['REQUEST_URI']))
+			if (isset($crumb['url']) && 
+			    str_replace('&amp;', '&', $crumb['url']) != basename($_SERVER['REQUEST_URI']))
 			  {
 			    echo '<a class="breadcrumbs" href="'.$crumb['url'];
 			    echo '">'.$crumb['title'];
